+++ attributes/default.rb
@@ -36,9 +36,6 @@ default['openstack']['network']['lock_path'] = '$state_path/lock'
 # Specify policy.json remote file to import
 default['openstack']['network']['policyfile_url'] = nil
 
-# DB Stamp comes late in dev cycle, allow override for development and test
-default['openstack']['network']['db_stamp'] = node['openstack']['release']
-
 # Gets set in the Network Endpoint when registering with Keystone
 default['openstack']['network']['region'] = node['openstack']['region']
 default['openstack']['network']['service_user'] = 'neutron'