+++ spec/server_spec.rb
@@ -14,8 +14,8 @@ describe 'openstack-network::server' do
 
     include_context 'neutron-stubs'
 
-    it 'uses release db stamp' do
-      expect(chef_run).to run_bash('migrate network database').with_code(/stamp juno/)
+    it 'uses db upgrade head' do
+      expect(chef_run).to run_bash('migrate network database').with_code(/upgrade head/)
     end
 
     it 'does not install neutron-server when nova networking' do